Create Python program that display the first N prime numbers, when user enters N numbers.
Write a Python function that will display a specific number of prime numbers for example if we call the function with 5 as a parameter then it will display five prime numbers like 1 2 3 5 and 7. Your program should ask a user to enter the number of prime numbers to be display for example if a user enter 7 then the program will call prime function ( prime(num) ) with parameter num = 7. The prime function will now display 7 prime numbers like 1 2 3 5 7 11 and 13.